changed salary to double type

+++ src/main/java/Employee.java
@@ -2,9 +2,9 @@ public abstract class Employee {
 
     private String name;
     private int niNumber;
-    private int salary;
+    private double salary;
 
-    public Employee(String name, int niNumber, int salary){
+    public Employee(String name, int niNumber, double salary){
         this.name = name;
         this.niNumber = niNumber;
         this.salary = salary;
@@ -19,7 +19,7 @@ public abstract class Employee {
         return this.niNumber;
     }
 
-    public int getSalary(){
+    public double getSalary(){
         return this.salary;
     }
 }
